Create Diagnostics component to be able to scale to support multiple diagnostics easily
