+++ src/freenet/node/diagnostics/diagnostics/Diagnostics.java
@@ -1,0 +1,5 @@
+package freenet.node.diagnostics.diagnostics;
+
+public interface Diagnostics {
+    void start();
+}
